--- x-pack/plugins/security_solution/cypress/downloads/rules_export.ndjson
@@ -1,4 +1,0 @@
-{"id":"6ee22d00-1c4d-11ee-bf9c-d7a1790ad38b","updated_at":"2023-07-06T22:35:40.414Z","updated_by":"system_indices_superuser","created_at":"2023-07-06T22:35:40.414Z","created_by":"system_indices_superuser","name":"Rule to export","tags":["test","newRule"],"interval":"100m","enabled":true,"revision":0,"description":"The new rule description.","risk_score":17,"severity":"high","note":"# test markdown","output_index":"","author":[],"false_positives":["False1","False2"],"from":"now-50000h","rule_id":"5cf25947-1824-4d0d-bac5-5d3099843722","max_signals":100,"risk_score_mapping":[],"severity_mapping":[],"threat":[{"framework":"MITRE ATT&CK","tactic":{"name":"Credential Access","id":"TA0006","reference":"https://attack.mitre.org/tactics/TA0006"},"technique":[{"id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","name":"OS Credential Dumping","subtechnique":[{"id":"T1003.008","name":"/etc/passwd and /etc/shadow","reference":"https://attack.mitre.org/techniques/T1003/008"}]},{"name":"OS Credential Dumping","id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","subtechnique":[]}]},{"framework":"MITRE ATT&CK","tactic":{"name":"Credential Access","id":"TA0006","reference":"https://attack.mitre.org/tactics/TA0006"},"technique":[{"id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","name":"OS Credential Dumping","subtechnique":[{"id":"T1003.008","name":"/etc/passwd and /etc/shadow","reference":"https://attack.mitre.org/techniques/T1003/008"}]}]}],"to":"now","references":["http://example.com/","https://example.com/"],"version":1,"exceptions_list":[],"immutable":false,"related_integrations":[],"required_fields":[],"setup":"","type":"query","language":"kuery","index":["apm-*-transaction*","auditbeat-*","endgame-*","filebeat-*","logs-*","packetbeat-*","traces-apm*","winlogbeat-*","-*elastic-cloud-logs-*"],"query":"host.name: *","actions":[]}
-{"id":"71479440-1c4d-11ee-bf9c-d7a1790ad38b","updated_at":"2023-07-06T22:35:43.411Z","updated_by":"system_indices_superuser","created_at":"2023-07-06T22:35:43.411Z","created_by":"system_indices_superuser","name":"rule with exceptions","tags":["test","newRule"],"interval":"100m","enabled":true,"revision":0,"description":"The new rule description.","risk_score":17,"severity":"high","note":"# test markdown","output_index":"","author":[],"false_positives":["False1","False2"],"from":"now-50000h","rule_id":"2","max_signals":100,"risk_score_mapping":[],"severity_mapping":[],"threat":[{"framework":"MITRE ATT&CK","tactic":{"name":"Credential Access","id":"TA0006","reference":"https://attack.mitre.org/tactics/TA0006"},"technique":[{"id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","name":"OS Credential Dumping","subtechnique":[{"id":"T1003.008","name":"/etc/passwd and /etc/shadow","reference":"https://attack.mitre.org/techniques/T1003/008"}]},{"name":"OS Credential Dumping","id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","subtechnique":[]}]},{"framework":"MITRE ATT&CK","tactic":{"name":"Credential Access","id":"TA0006","reference":"https://attack.mitre.org/tactics/TA0006"},"technique":[{"id":"T1003","reference":"https://attack.mitre.org/techniques/T1003","name":"OS Credential Dumping","subtechnique":[{"id":"T1003.008","name":"/etc/passwd and /etc/shadow","reference":"https://attack.mitre.org/techniques/T1003/008"}]}]}],"to":"now","references":["http://example.com/","https://example.com/"],"version":1,"exceptions_list":[{"id":"70b20b00-1c4d-11ee-bf9c-d7a1790ad38b","list_id":"test_exception_list","type":"detection","namespace_type":"single"}],"immutable":false,"related_integrations":[],"required_fields":[],"setup":"","type":"query","language":"kuery","index":["apm-*-transaction*","auditbeat-*","endgame-*","filebeat-*","logs-*","packetbeat-*","traces-apm*","winlogbeat-*","-*elastic-cloud-logs-*"],"query":"host.name: *","actions":[]}
-{"_version":"WzE0OCwxXQ==","created_at":"2023-07-06T22:35:42.384Z","created_by":"system_indices_superuser","description":"Test exception list description","id":"70b20b00-1c4d-11ee-bf9c-d7a1790ad38b","immutable":false,"list_id":"test_exception_list","name":"Test exception list","namespace_type":"single","os_types":[],"tags":[],"tie_breaker_id":"7e7a7f37-7ed3-49b8-827f-6add0a84400b","type":"detection","updated_at":"2023-07-06T22:35:42.384Z","updated_by":"system_indices_superuser","version":1}
-{"exported_count":3,"exported_rules_count":2,"missing_rules":[{"rule_id":"rule_3"},{"rule_id":"rule_6"},{"rule_id":"rule_4"},{"rule_id":"rule_1"},{"rule_id":"rule_7"},{"rule_id":"rule_2"},{"rule_id":"rule_5"}],"missing_rules_count":7,"exported_exception_list_count":1,"exported_exception_list_item_count":0,"missing_exception_list_item_count":0,"missing_exception_list_items":[],"missing_exception_lists":[],"missing_exception_lists_count":0,"exported_action_connector_count":0,"missing_action_connection_count":0,"missing_action_connections":[],"excluded_action_connection_count":0,"excluded_action_connections":[]}
